[fidl] Remove clang-format column limit for fidlgen
To allow our fidlgen templates to be optimized for maintainance rather
than generating readable output we pass generated files through a
language-specifc formatter. For C++ this is clang-format. Unfortunately
clang-format can be slow for large, complex files.

It turns out that one way to mitigate this is to remove the column
limit, passing "ColumnLimit: 0" as a configuration option.
Counterintuitively this doesn't actually result in crazy long lines,
just arguably mildly worse line breaking choices, though that's
subjective. Look at the changed goldens in this commit to see the impact
on the output.

The upside is a performance win. On my workstation, the time to format
all of the goldens drops from 15s to 5s with a modest drop in maximum
memory usage from 72M to 68M.

Formatting the LLCPP bindings for fuchsia.posix.socket - our largest
generated code (that is excluded from formatting for performance
reasons) runtime drops from 41s to 21s but max RSS for both was about
1.2G.

I don't think this will let us stop using a size limit for formatting
but at least it should speed things up

The clang-format args used to be scattered across a Go file and some
BUILD.gn files but now they're just all in the generator.go, making it
simple to experiment with tweaks in the future.

Bug: 78303
Bug: 49757
Change-Id: I53316832aee802aafc09d09f4e60b0963fae87bd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680908
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>