[docs][lsan] Add a short description of how LeakSanitizer works
I thought this would be useful to add, because otherwise the
information is three steps away from this page, via these links:

https://clang.llvm.org/docs/LeakSanitizer.html ->
https://github.com/google/sanitizers/wiki/AddressSanitizerLeakSanitizer ->
https://github.com/google/sanitizers/wiki/AddressSanitizerLeakSanitizerDesignDocument

Some other leak checker tools have relied on programs freeing
allocations before running the leak checker and exiting.  It's easy
for there to be misconceptions that LeakSanitizer has similar
requirements, and that allocations that are never freed (or threads
that never exit) are considered leaks, which is not necessarily the
case with LeakSanitizer.

Change-Id: Iaea5013b102e74a1796cf2cb1b247b07467961b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/679995
Reviewed-by: Leonard Chan <leonardchan@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>