[kernel][scheduler] Explicitly disable preemption during profile changes.
Explicitly disable preemption while holding the thread lock during a
thread base profile change. See the comment left in the CL for the
reasoning here.

This CL is a quick-fix for an issue discovered via a flaking test.
Moving forward, the plan is to add some static analysis tools to
ensure that there are no other code paths (now or in the future) which
could accidentally make the same mistake.

Bug: 100625
Change-Id: I860d90cecca352420024fedb8ea8d30a77664ede
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681915
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>