[ffx][repo] deregister should not talk to target
Previously, `ffx repository deregister` has the semantics where when a
repository was deregistered from a target, the target-side configuration
would also be removed. However, this doesn't match the workflows of `fx
serve` and `fserve`, where killing those processes leaves behind the
target-side configurations.

In order to ease our migration to `ffx repository`, this change modifies
the ffx repository server to stop trying to clean up the target.
Instead, we only clean up our internal tracking of this registration,
and will stop trying to re-register our repository on start, or when we
first connect to a device. This then has the side benefit that we can
remove a registration if the target is offline.

As part of cleaning this up, this also allows us to deregister
repositories, even if the repository doesn't exist anymore. This isn't
supposed to happen, but users can run into it if they are directly
manipulating the ffx configuration.

In order to do this, this changes the ffx context object to expose a way
to query for target information, even if the target is offline.

Fixed: 98789
Change-Id: I79ef96692430c7ae9b58711607743fda0f171f2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680284
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Erick Tryzelaar <etryzelaar@google.com>
Reviewed-by: Ben Keller <galbanum@google.com>