[audio][clock] Add RecoveredClock
This CL adds a class which can be used to recover a clock from a stream
of position updates (e.g. from a device driver). This is being factored
out of AudioDriver and AudioClock. The behavior should be identical to a
a DeviceAdjustable AudioClock as used in the following code:

https://cs.opensource.google/fuchsia/fuchsia/+/main:src/media/audio/audio_core/audio_driver.cc;drc=73cfbf4538bbcec51bf87420803ca2da65060f8d;l=588
https://cs.opensource.google/fuchsia/fuchsia/+/main:src/media/audio/lib/clock/audio_clock.cc;drc=a05c6d9f6f02df7d46611e8b2626e9ad96cd4a1f;l=318

Bug: 87651
Change-Id: I3158c27c4e5f2a2f877b19793c1f296ee7240672
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680506
Reviewed-by: Alper Gungormusler <alperg@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>
Commit-Queue: Tom Bergan <tombergan@google.com>