[board][nelson] Add a separate GPIO device for bank H GPIOs
Add a GPIO device that only exposes GPIOs from bank H. This bank uses
separate registers from all other GPIOs, so no synchronization is
required between this and the main GPIO instance.

The exception to this is the radar interrupt gpio -- the GPIO H device
will not be able to provide interrupts for GPIOs it exposes, so this
GPIO must be exposed by the main device. The interrupt GPIO is only
configured once during init, making synchronization in the Selina driver
trivial.

By being a platform device rather than a protocol implementation device,
this GPIO driver instance can be colocated with the SPI and Selina
drivers in order to reduce the rate of errors in long-running radar
tests.

Bug: 95139
Bug: b/216348080
Test: device-enumeration-test passed on Nelson
Test: Ran radarutil on Nelson
Change-Id: Iae3006273870e287ba9595e8bf202449225226e7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/678131
Reviewed-by: Christopher Anderson <cja@google.com>
Reviewed-by: Andres Oportus <andresoportus@google.com>
Commit-Queue: Braden Kell <bradenkell@google.com>