[kernel][vm] Simplify PrepareForWriteLocked exit conditions
PrepareForWriteLocked succeeds if *any* dirty pages can be found at the
start of the specified range, even if pages further down the range might
required a DIRTY request later. Let the lookup succeed with the dirty
pages we have accumulated so far, and come back in again for the DIRTY
request later. This unifies the logic with the non-TRAP_DIRTY case which
can also succeed with only some of the pages dirty at the start of the
range. It also makes it easier to reason about the value of
|dirty_len_out| returned in success and failure cases.

Also cleanup some internal page list traversals by unifying the page
accumulation logic.

Test: core-pager-writeback

Bug: 63989
Run-All-Tests: true
Change-Id: I69ab15cde8fe2cf7412cfa7776ecd86ab86bfdd5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681449
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>