[netstack3] Bind unbound socket on sendto() call
Handle an application calling sendto() on an unbound socket by first
binding it to the unspecified address and then sending a packet on it.
This behavior is consistent with Linux, as determined through
experimentation: if the send fails, i.e. because the address is
unreachable, the socket is still bound to a local address afterwards.

Bug: 100939
Change-Id: I91811a7728137ee1f76ff471aeb4203cb7073117
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680943
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Fuchsia-Auto-Submit: Alex Konradi <akonradi@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>