[fidl][gidl] Test handle closure during failures
This tests a case that was encountered when building the new LLCPP
encoder and decoder. It is important to ensure the handles are
closed regardless of the location of the error and that the
resulting handle array is properly sized.

Bug: 100931
Bug: 96626
Bug: 100945
Change-Id: I345be56e54ab9eb30804fe1f714b78eba5899514
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681788
Fuchsia-Auto-Submit: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Mitchell Kember <mkember@google.com>