[artifactory] Stop uploading Jiri snapshot
The snapshot is already uploaded by the recipe, which is the more
"proper" place to be doing this (during the checkout phase), not
post-build.

The reason this was added was because users had requested that the
snapshot show up in fuchsia-artifacts buckets, but fxbug.dev/51741
migrated the Jiri snapshot upload to fuchsia-artifacts, so this is
always a no-op for artifactory anyway. (It just gets deduplicated every
time).

Bug: 51741
Change-Id: Ic67c5db7f10999f4872350482aacda94b6b6372e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682543
Fuchsia-Auto-Submit: Anthony Fandrianto <atyfto@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Commit-Queue: Anthony Fandrianto <atyfto@google.com>