[artifactory] Move nonexistence checks earlier in execution
Artifactory already filters out nonexistent files, but only did this at
upload time. This change modifies the execution to filter prior to
emitting the manifest, so that the manifest is not polluted with files
that do not exist.

This allows gcs-util to enforce that every file in the input manifest
does exist, which is a good assertion to have. Each project should take
responsibility over emitting a manifest with entirely valid references.

Bug: 92697
Change-Id: I5d43534bdae82904eb93b5401e315cbbf60cff62
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682802
Reviewed-by: Ina Huh <ihuh@google.com>
Fuchsia-Auto-Submit: Anthony Fandrianto <atyfto@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>