[component_manager] Purge invokes Destroy
This change causes the Purge action to automatically invoke Destroy.
Component instance destruction paths are changed to schedule the Purge
action and synchronously wait on it.

This is the minimal change to get the behavior proposed in
fxbug.dev/100652. It is preparation toward merging replacing Purge
entirely with Destroy, which will happen in a followup.

Bug: 100652, 89503

Change-Id: I727a22827eb0720989deedcad96927ed9b5c6555
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/677672
Reviewed-by: Justin Mattson <jmatt@google.com>
Reviewed-by: Xyan Bhatnagar <xbhatnag@google.com>
Commit-Queue: Gary Bressler <geb@google.com>