[bt][core] Remove unused production v1 components
bt-gap and bt-init are now CFv2 in production, so we can remove the v1
components. For bt-gap, this just means removing the v1 manifest and
build targets. For bt-init, the v1 component was actually a separate
binary with a different main.rs, so remove that as well as the manifest/
build targets.

Bug: 89297
Test: `fx test //src/connectivity/bluetooth` to verify nothing broke.
Change-Id: I47f1ba038339b80cc58f30f240c2b31c4ff7a72c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680571
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>
Reviewed-by: Ani Ramakrishnan <aniramakri@google.com>