[test_manager] Creates a specific type of test for starnix
The starnix test currently run a linux binary with a starnix runner
outside of the test realm.

This is problematic for 2 reasons:
- The test is not hermetic at all.
- There is no code coverage on the runner code

To fix this, the runner will instead be run inside the realm of the
test.

The runner requires a number of permission that we do not want to give
to every test. To allow only giving these permissions to the test that
needs it, this CL introduce a specific type of tests and a specific
collection to run these tests in.

Change-Id: If1855dc02d5c9406d1ebc331ed6a7aa6c9cce87c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682964
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>