[testrunner] Move test outputs after recording test result.
In http://fxr/680306, we used a temp dir for the test outputs dir which
some tests write to, and then moved those contents back to the
originally designated output dir. This broke tests which use the
testparser to store paths to those outputs as outputfiles of the test in
the test result because the paths would point to the tmp dir that no
longer existed. This change allows testparser to access the paths in the
tmp dir and record the output files which will then be moved (and paths
updated) in the call to outputs.Record(). After the output files are no
longer pointing to paths in the tmp dir, we can then move the rest of
the contents to the original test output directory so that they can be
uploaded with the swarming task outputs.

Bug: 100948
Change-Id: I4df077c38217ac2f91ce476db5f5964dbb7f3f0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682023
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>