[build] Delete sdk_package
This template is broken because it attempts to use fuchsia_package() on
the host toolchain. fuchsia_package() cannot be used with anythign but
the target toolchain, and when it does it produces an empty group.
Delete this template and all of its uses, which were all unfortunately
no-ops.

This was written and reviewed by Misha and Sean, both of whom have
unfortunately left the team.

Bug: 100612
Change-Id: I756cdc33a207ccd41bfde643e015b1161b7d4135
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/680933
Reviewed-by: Aaron Wood <aaronwood@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>