[installer] Log FVM write status.
The FVM write takes a long time. In some instances the installer tests
flake, and it's not clear if it's hanging or just taking a while to
write the FVM.

This CL adds some progress logging so we can get an idea of how
quickly/slowly the installer is progressing during the FVM write.

Bug: 100712
Change-Id: I76c1fb67eafe45319d682da2ccfabbe8ff16f454
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683363
Commit-Queue: Simon Shields <simonshields@google.com>
Reviewed-by: Sarah Pham <smpham@google.com>