[scripts/owner] Find owners for projects in path
Extend `suggest_owners.py` script to accept a `--path` argument instead
of `--jiri-manifest`, to find owners for the project(s) in the given the
given path.

The `--path` argument accepts wildcard and multiple arguments.

For now, the script only works for projects that use gn, and will not
return owners for projects that use other build systems.

Drive-by: fix parsing of OWNERS files that include another OWNERS file.

Bug: none

Change-Id: I499f9ec93ebe471d2990b92a3bab675697fcfc40
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682946
Commit-Queue: Marina Ciocea <marinaciocea@google.com>
Reviewed-by: Chris Palmer <palmer@google.com>
Reviewed-by: Shai Barack <shayba@google.com>