[cleanup] remove all stale suppressions in zircon
bug: 58162

This change is part of the C++ implicit conversions project. These are the build files that compile without more changes. The purpose is to shorten the list of stale suppressions in the kernel to find which ones still need to be fixed.

Change-Id: Ie7daa7d3b2d3cc7e2bd9ea017d2037423f5d7a84
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683422
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Fuchsia-Auto-Submit: Shai Barack <shayba@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>