[minfs] remove "paranoid mode"
This didn't seem to be used too much. It was a commented out #define
that could add a check that we zero'd out the bits between the end of a
file and the end of the block alignment.

Bug: 100846
Change-Id: Ide751ba1905e5f520580b14fd3f29ff868c7c859
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683137
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Stephen Demos <sdemos@google.com>
Reviewed-by: Aaron Drew <ripper@google.com>