[pkg] Use narrowed fuchsia-pkg URL types
PkgUrl could represent URLs with/without a resource (indicating a
component or just a package) and URLs with/without a hash
(pinned or unpinned).

AbsolutePackageUrl only represents URLs without a resource (but still
supports optional hash), and the Pinned and Unpinned versions require
that the URL additionally have or not have a hash.

Bug: 99960
Change-Id: I5e8db273964213b9d4efea078468c1ecc2aa0ea9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682669
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Simon Shields <simonshields@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>