[tools] Trim down unnecessary cross-compiling
The infrastructure never runs x64 target tests from an arm64 host, so we
don't need to build for an arm64 host when the target arch is x64.

Additionally, testrunner is the only tool from `//tools/testing:tools`
that actually needs to be built for both toolchains in arm64 target
builds. The other tools either always run on an x64 host (resultdb,
tefmocheck) or aren't used directly by infrastructure (testparser is
only used as a library in testrunner, and seriallistener is now only
used for ZBI tests, which include it as a dep).

Most notably, this will stop building ffx for arm64 on x64 builders. ffx
takes up to 8 minutes to build, so this should be a significant win.

Change-Id: I86e69dc61edd4e75fa8ddd808daafa3469e84272
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683649
Commit-Queue: Oliver Newman <olivernewman@google.com>
Reviewed-by: Ina Huh <ihuh@google.com>