[identity] Remove scenario from the account API.
The idea with a threat scenario several years ago is authenticators
would be able to assess authentication quality against several attackers
in parallel, avoiding the need to lock a product's assumption about its
threat model into the implementation.

In today's world we're no longer pursuing cross-product identity
consistency and its inevitable that products will want to bake their
assumptions about the threat model into the set of authenticators they
enable and the policy they enforce. This means we really dont need
scenario as an axis in the API surface any more. Things get quite a bit
simpler for all components concerned when we remove it.

Note that the removal of the "we don't know the scenario" in the listener
means we no longer need a separate `InitialAccountState` struct, but
removing that also means we need to publish a hardcoded auth state in
the initial listener implementation.

Also note that `AccountManager.GetAccountAuthStates` is a lot less
useful without a scenario so I've removed it. Interested parties could
still register a listener.

Bug: 100785
Change-Id: Ia8bb447c476929f941c3f599662b02ba37e81520
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682584
Commit-Queue: Jody Sankey <jsankey@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>