[identity] Simplify auth state in the account API.
The previous auth state was designed around the ambient concerns of user
presence and engagement. There are two reasons this isn't relevant for
today's situation:
1. We're no longer persuing cross-product identity consistency so there
   isn't a need for a "one-size-fits-all" cross-product description of
   authentication state.
2. We're not planning to build any ambient authentication mechanisms in
   the foreseeable future, and its those that would be able to actually
   distinguish between presence and engagement.

This CL strips out the presence and engagement fields and alters the
summary enum to something a lot more prosiac and similar to existing
products. The details of policy for if and when a particular summary
state (e.g. "recently authenticated") is used fall to the product.

We also move to a flexible enum and tables so that the auth
state can be expanded in the future.

Bug: 100785
Change-Id: I290b59aa4589dac5836834dceac0cf775b9ef134
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/682763
Commit-Queue: Jody Sankey <jsankey@google.com>
Reviewed-by: Rebecca Silberstein <silberst@google.com>
Reviewed-by: Satsuki Ueno <satsukiu@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>