[sdk] Remove support for fsys2.Events from Dart Realm Builder
Dart Realm Builder will be published to the SDK, but Events (under
fuchsia.fsys2) are not supported in the SDK yet, so they must be
removed. Events are coming to the Fuchsia SDK in the near future, and
this functionality can be restored at that time.

Change-Id: I2dd04437048bb69d43bd202975569f86c351cf8e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683851
Commit-Queue: Rich Kadel <richkadel@google.com>
Reviewed-by: Alexander Biggs <akbiggs@google.com>