[power] Move throttling metrics to PlatformMetrics
Moves the code and logic for tracking throttling metrics out of
ThermalPolicy and into PlatformMetrics. This refactor prepares for a
subsequent change where ThermalStateHandler will be instrumented to
provide throttling metrics instead of ThermalPolicy.

In the process, this change removes support for the
`thermal_limiting_elapsed_time` Cobalt metric. The Power Manager
currently dispatches data for this event. However, we've never used this
data in its lifetime. Now that the metric is expired in Cobalt, it's a
good time to just remove it.

This change temporarily removes the `available_power_hist` and
`cpu_power_usage` properties from being collected in ThermalPolicy. They
will be re-added in a subsequent CL to improve their usefulness
(fxbug.dev/98245).

Bug: 86871
Bug: 98245
Test: fx test power-manager-tests

Change-Id: I82b3419edbc1a5248984df9fa1907855cccecb71
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/670242
Reviewed-by: Didi She <didis@google.com>
Commit-Queue: Patrick Shickel <pshickel@google.com>