[audio][drivers] Fix audio-driver-ctl connection check
Rely only on not being able to communicate with a driver
(getting formats) to find out if a driver is present at
a particular devfs location. This because we use a feed
forward API which may not return an error when setting up
a connection.
fxrev.dev/678038 checked for stream channel peer closed
which fails depending on timing since it takes some time
for the signal to get asserted, removed the bad check.

Test: audio-driver-ctl info on Nelson shows driver info

Change-Id: I1c9174a744325e5d6a0deefe7baafeb7b27786f6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684163
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>