[rust] Use the path from the checkout for llvm-profdata-rust.
The rust version of llvm-profdata links LLVM as a shared library so we'd
need the shared library file together with the llvm-profdata binary in
order to make the tool work. We can possibly upload both with the
orchestration inputs so that the coverage processor builder can use it,
but that would require some more refactoring to be able to preserve the
relative paths and somehow know the path to the llvm-profdata tool
within the rust prebuilts dir. On the other hand, the coverage processor
builder already does a checkout from which it could access the tool
directly, getting the path through the tool_paths.json.

Bug: 95345
Change-Id: Iff5ec00343b4afed68eb879e082f0a60190d1576
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683869
Fuchsia-Auto-Submit: Ina Huh <ihuh@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Tyler Mandry <tmandry@google.com>