[fxfs] Avoid validating checksums of extents marked as deleted.
This skips checksum validation if an owner object_id was marked as
deleted prior to the "valid_to" offset in the journal file.

I suspect it also fixes a bug whereby if we have more than one corrupt
journal entry, we would replay up to the last corruption because the
break statement during validation was only breaking out of one layer of
for loop.

Bug:99146

Change-Id: I8e307a9fdecf75e7353a138da87e44662a8df0d2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681411
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Aaron Drew <ripper@google.com>