[media/vnext] add vnext to the build and fix build errors
This is a reland of fxr/683664. transport_controller_test.cc has
been fixed and Run-All-Tests added to ensure that the tests are run.

This CL adds //src/media/vnext to the //src/media group and
//src/media/vnext::tests to //src/media:tests. It also fixes some
build errors that cropped up, because vnext was not getting built
by the trybots.

Test: build only - no behvior change

Run-All-Tests: true

Change-Id: I82e0eca00779b9631ddabc7f9e4a50d51ff79d3c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684217
Commit-Queue: Dale Sather <dalesat@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>