[audio_core] Reboot audio_core on terminate
A version of audio_core built out-of-tree is currently labeled
reboot-on-terminate. The original motivation was that some code
(including in-tree code) is (currently) not designed to correctly
reconnect to audio_core and reconstruct audio state after an audio_core
crash, so it was deemed safer to just reboot the system.

The in-tree audio_core was never labeled reboot-on-terminate, which was
simply an oversight. This CL corrects that oversight.

Change-Id: I2a344178ce15f52936ce2e663df0c6b84687a798
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683883
Commit-Queue: Tom Bergan <tombergan@google.com>
Reviewed-by: Drew Fisher <zarvox@google.com>
Reviewed-by: Martin Puryear <mpuryear@google.com>