[libabr] fix magic constant definition
Putting a static const variable in a header actually tells each
translation unit to make its own copy of the variable, rather than
sharing a single copy.

This can cause a (small) duplication, but more importantly, files
which include this header but *don't* use this particular constant
will trigger an unused variable compiler error.

This CL just switches it to be a macro instead. The "right" way to do
this is to define an extern char* in the header and provide the
definition in a .c file, but since we also want to declare the length
constant immediately after it's best to keep the definitions
side-by-side.

Change-Id: I0174e4b0ead5375074013c99b04f80f18cf4be0b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684362
Reviewed-by: Yecheng Zhao <zyecheng@google.com>
Commit-Queue: David Pursell <dpursell@google.com>