[ally] Optimize semantic tree updates.
This change reduces lookups into nodes_to_be_updated_ by re-using
the iterator returned from the .find() call when it is valid. It
also removes a duplicate call to update.TakeDeleteNodeId(), which
reduces some overhead and more closely fits with the concept of
"taking". This also  prevents a double std::move(), which is technically
okay in this case since the type being moved is uint32_t, but should
still be avoided.

Change-Id: I5123b3fcd3a0c663d166355a18276536820480d9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684170
Reviewed-by: Lucas Radaelli <lucasradaelli@google.com>
Commit-Queue: Nick Cano <nickcano@google.com>
Reviewed-by: Alexander Brusher <abrusher@google.com>