[mdns] make address fixups safe
This CL changes the way messages are distributed for transmission.

In the mdns service, a single MdnsTransceiver object is responsible
for sending and receiving DNS messages across all usable interfaces
over both IPv4 and IPv6. To accomplish this, it aggregates any number
of MdnsInterfaceTransceiver objects, each of which is responsible for
I/O through a single interface over IPv4 or IPv6 (not both). Often,
sent messages are multicast, meaning they are sent through multiple
interface transceivers.

When address resources (A and AAAA records) need to be sent, those
resources need to be interface-specific. Specifically, they need to
contain the addresses of interface itself. To support this, there is
the notion of an 'address placeholder' resource, which is just an A
resource with an invalid IP address.

Each interface transceiver, therefore, needs to 'fix up' the message
before sending it, replacing address placeholders with one or two
address resources for the interface. When a message needs to be
sent through multiple interface transceivers, the message is
distributed sequentially to one transceiver at time.

Previously, each interface transceiver was allowed to mutate the
message, so the next interface transceiver would be asked to send
the mutated message. This worked fine when *all* addresses were
fixed up (rather than just placeholder addresses). Recently, we
started to send address resources that were not placeholders. The
fixup of non-placeholder addresses was suppressed, but this mutation
issue was forgotten (I forgot it). As a result, interface addresses
were being propagated from one transceiver to the next, so they would
show up on the wrong links.

This CL changes the SendMessage methods so the message is now
immutable. Fixups involve creating a new message, but the resources
they contain aren't copied (they are managed with shared_ptr) so this
is not terribly expensive.

Test: supplemented existing netemul test - fails without fix

Change-Id: Ie7e8b92e9a8035fcefcb7b44b262a93eb763bd81
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/681544
Commit-Queue: Dale Sather <dalesat@google.com>
Fuchsia-Auto-Submit: Dale Sather <dalesat@google.com>
Reviewed-by: Nate Hurley <nahurley@google.com>
(cherry picked from commit ca97b81a85305e4ed68878d6effe33828597599e)