[fshost][test] Run fshost tests with Fxfs
Templatizes the fshost integration test suite so we can run the test
with various fshost configurations.  This permits running the test suite
with Fxfs, which is also added by this CL.  (f2fs can be added in a
follow-up; it was not turnkey so I left it out of this CL.)

A previous attempt to run the fshost integration tests with varied
configuration was fxrev.dev/679445, which used Gtest parametrized tests
and RealmBuilder, did not allow variance in the component manifest
files, which would cause problems in the future.  This approach is less
tidy but allows both the structured config and the component manifest to
be varied.

Fixed: 99714
Change-Id: I4e24eb702c00e84ae8b4e3a833dd128924670891
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683026
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>