[vfs] Remove Journal OperationTracker
Replace the Journal OperationTracker with the equivalent data structure.

Discovered as part of metrics cleanup, removed to reduce confusion with
other metrics/statistics-related trackers. The only functionality that
OperationTracker provided which the backing IntervalTree does not is
the Remove function, which was not used in production (only tests).

Remove can be implemented as a free function if required in the future.

Test: No functional change.
Change-Id: Ib5de85b8995ba3b1d58d2115b8426c85b733d252
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/683629
Fuchsia-Auto-Submit: Brandon Castellano <bcastell@google.com>
Reviewed-by: Stephen Demos <sdemos@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>