[kernel][vm] Retain dirty pages on detach
Only clean pages are decommitted when a VMO is detached from the pager.
The user pager is expected to query dirty pages on detach and write them
back so that they can be marked clean. Once marked clean, the kernel is
then free to evict them. Previously, all of the VMO's pages would be
dropped on detach, which was meant to serve as a deterministic failure
mode for blobfs. This behavior remains unchanged as blobfs only uses
read-only VMOs, so all pages are clean and would be decommitted still.

Also add some tests that verify various behaviors related to detach.

Test: core-pager-writeback
Bug: 63989
Run-All-Tests: true
Change-Id: Id2da4c5465089bb72352aaf66cda8397c8217ba3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/684369
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>